pyproject.toml
[tool.poetry]
name = "dj-backend"
version = "0.1.0"
description = ""
authors = ["Ilya Filimonov <Filimonov_IS@mail.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.3"
django-extensions = "3.1.3"
django-configurations = "^2.2"
Pillow = "^8.2.0"
graphene = "^2.1.8"
graphene-django = "^2.15.0"
django-graphql-jwt = "0.3.1"

[tool.poetry.dev-dependencies]
poethepoet = "^0.10.0"

[tool.poe.tasks]
serve = "python manage.py runserver"
superuser = "python manage.py createsuperuser"
flush = "python manage.py flush"
secret = "python -c 'from django.core.management.utils import get_random_secret_key; print(get_random_secret_key())'"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

